UdfpsAnimations: Add 'none' to the styles Since we switched to single Settings.UDFPS_ANIM_STYLE , we have to add an item to the arrays that don't have any drawables in the index which we consider as disabled value for the Settings.
we consider value 0 means disabled.

Evolution X notes:

Since we ship WallpaperPickerGoogleRelease, we will not be
using the new picker impl. Instead we only remove the
MainSwitchBar from UdfpsAnimation picker frag of the Evolver
and add 'none' here with a single transparent drawable.

This single transparent drawable will not be drawn in SystemUI
and only within the UdfpsAnimation picker frag since index 0
is disabled.

Change-Id: I00a9f8847f3f596705f48b8bf8a5291ebf3c9776